+++ Tests/RethinkDBTests/MathAndLogicTests.swift
@@ -13,31 +13,31 @@ class MathAndLogicTests: BaseTests {
     
     func testAdd() throws {
         let conn = try r.connect(host: Tests.host)
-        let result: Int = try (r.expr(2) + 2).run(conn)
+        let result: Int64 = try (r.expr(2) + 2).run(conn)
         XCTAssert(result == 4, "Expected result to equal 4, found \(result)")
     }
     
     func testSubtract() throws {
         let conn = try r.connect(host: Tests.host)
-        let result: Int = try (r.expr(4) - 2).run(conn)
+        let result: Int64 = try (r.expr(4) - 2).run(conn)
         XCTAssert(result == 2, "Expected result to equal 2, found \(result)")
     }
     
     func testMultiply() throws {
         let conn = try r.connect(host: Tests.host)
-        let result: Int = try (r.expr(2) * 2).run(conn)
+        let result: Int64 = try (r.expr(2) * 2).run(conn)
         XCTAssert(result == 4, "Expected result to equal 4, found \(result)")
     }
     
     func testDivide() throws {
         let conn = try r.connect(host: Tests.host)
-        let result: Int = try (r.expr(8) / 2).run(conn)
+        let result: Int64 = try (r.expr(8) / 2).run(conn)
         XCTAssert(result == 4, "Expected result to equal 4, found \(result)")
     }
     
     func testMod() throws {
         let conn = try r.connect(host: Tests.host)
-        let result: Int = try (r.expr(10) % 3).run(conn)
+        let result: Int64 = try (r.expr(10) % 3).run(conn)
         XCTAssert(result == 1, "Expected result to equal 1, found \(result)")
     }
